rename new arrow from SFM to MVG
