+++ arrows/mvg/algo/initialize_cameras_landmarks.h
@@ -30,26 +30,26 @@
 
 /**
  * \file
- * \brief Header for SFM camera and landmark initialization algorithm
+ * \brief Header for MVG camera and landmark initialization algorithm
  */
 
-#ifndef KWIVER_ARROWS_SFM_INITIALIZE_CAMERAS_LANDMARKS_H_
-#define KWIVER_ARROWS_SFM_INITIALIZE_CAMERAS_LANDMARKS_H_
+#ifndef KWIVER_ARROWS_MVG_INITIALIZE_CAMERAS_LANDMARKS_H_
+#define KWIVER_ARROWS_MVG_INITIALIZE_CAMERAS_LANDMARKS_H_
 
-#include <arrows/sfm/kwiver_algo_sfm_export.h>
+#include <arrows/mvg/kwiver_algo_mvg_export.h>
 
 #include <vital/algo/initialize_cameras_landmarks.h>
 
 namespace kwiver {
 namespace arrows {
-namespace sfm {
+namespace mvg {
 
 /// A class for initialization of cameras and landmarks
-class KWIVER_ALGO_SFM_EXPORT initialize_cameras_landmarks
+class KWIVER_ALGO_MVG_EXPORT initialize_cameras_landmarks
 : public vital::algo::initialize_cameras_landmarks
 {
 public:
-  PLUGIN_INFO( "sfm",
+  PLUGIN_INFO( "mvg",
                "Run SfM to estimate new cameras and landmarks "
                "using feature tracks." )
 
@@ -92,7 +92,7 @@ private:
   const std::unique_ptr<priv> m_priv;
 };
 
-} // end namespace sfm
+} // end namespace mvg
 } // end namespace arrows
 } // end namespace kwiver
