+++ arrows/mvg/register_algorithms.cxx
@@ -30,49 +30,49 @@
 
 /**
  * \file
- * \brief Plugin algorithm registration for SFM Arrow
+ * \brief Plugin algorithm registration for MVG Arrow
  */
 
-#include <arrows/sfm/kwiver_algo_sfm_plugin_export.h>
+#include <arrows/mvg/kwiver_algo_mvg_plugin_export.h>
 
 #include <vital/algo/algorithm_factory.h>
 
-#include <arrows/sfm/algo/initialize_cameras_landmarks.h>
+#include <arrows/mvg/algo/initialize_cameras_landmarks.h>
 
-// TODO: These are files that should move to SFM from Core
-//#include <arrows/sfm/close_loops_bad_frames_only.h>
-//#include <arrows/sfm/close_loops_appearance_indexed.h>
-//#include <arrows/sfm/close_loops_exhaustive.h>
-//#include <arrows/sfm/close_loops_keyframe.h>
-//#include <arrows/sfm/close_loops_multi_method.h>
-//#include <arrows/sfm/compute_ref_homography_core.h>
-//#include <arrows/sfm/estimate_canonical_transform.h>
-//#include <arrows/sfm/feature_descriptor_io.h>
-//#include <arrows/sfm/filter_features_magnitude.h>
-//#include <arrows/sfm/filter_features_scale.h>
-//#include <arrows/sfm/filter_tracks.h>
-//#include <arrows/sfm/hierarchical_bundle_adjust.h>
-//#include <arrows/sfm/initialize_object_tracks_threshold.h>
-//#include <arrows/sfm/keyframe_selector_basic.h>
-//#include <arrows/sfm/match_features_fundamental_matrix.h>
-//#include <arrows/sfm/match_features_homography.h>
-//#include <arrows/sfm/track_features_augment_keyframes.h>
-//#include <arrows/sfm/track_features_core.h>
-//#include <arrows/sfm/triangulate_landmarks.h>
+// TODO: These are files that should move to MVG from Core
+//#include <arrows/mvg/close_loops_bad_frames_only.h>
+//#include <arrows/mvg/close_loops_appearance_indexed.h>
+//#include <arrows/mvg/close_loops_exhaustive.h>
+//#include <arrows/mvg/close_loops_keyframe.h>
+//#include <arrows/mvg/close_loops_multi_method.h>
+//#include <arrows/mvg/compute_ref_homography_core.h>
+//#include <arrows/mvg/estimate_canonical_transform.h>
+//#include <arrows/mvg/feature_descriptor_io.h>
+//#include <arrows/mvg/filter_features_magnitude.h>
+//#include <arrows/mvg/filter_features_scale.h>
+//#include <arrows/mvg/filter_tracks.h>
+//#include <arrows/mvg/hierarchical_bundle_adjust.h>
+//#include <arrows/mvg/initialize_object_tracks_threshold.h>
+//#include <arrows/mvg/keyframe_selector_basic.h>
+//#include <arrows/mvg/match_features_fundamental_matrix.h>
+//#include <arrows/mvg/match_features_homography.h>
+//#include <arrows/mvg/track_features_augment_keyframes.h>
+//#include <arrows/mvg/track_features_core.h>
+//#include <arrows/mvg/triangulate_landmarks.h>
 
 
 namespace kwiver {
 namespace arrows {
-namespace sfm {
+namespace mvg {
 
 
 // ----------------------------------------------------------------------------
 extern "C"
-KWIVER_ALGO_SFM_PLUGIN_EXPORT
+KWIVER_ALGO_MVG_PLUGIN_EXPORT
 void
 register_factories( kwiver::vital::plugin_loader& vpm )
 {
-  kwiver::vital::algorithm_registrar reg( vpm, "arrows.sfm" );
+  kwiver::vital::algorithm_registrar reg( vpm, "arrows.mvg" );
 
   if (reg.is_module_loaded())
   {
@@ -84,6 +84,6 @@ register_factories( kwiver::vital::plugin_loader& vpm )
   reg.mark_module_as_loaded();
 }
 
-} // end namespace sfm
+} // end namespace mvg
 } // end namespace arrows
 } // end namespace kwiver
