+++ arrows/mvg/sfm_utils.h
@@ -33,15 +33,15 @@
 * \brief Header for utility functions for structure from motion
 */
 
-#ifndef KWIVER_ARROWS_SFM_SFM_UTILS_H_
-#define KWIVER_ARROWS_SFM_SFM_UTILS_H_
+#ifndef KWIVER_ARROWS_MVG_SFM_UTILS_H_
+#define KWIVER_ARROWS_MVG_SFM_UTILS_H_
 
 #include <vector>
 #include <unordered_set>
 
 #include <vital/vital_config.h>
 #include <vital/vital_types.h>
-#include <arrows/sfm/kwiver_algo_sfm_export.h>
+#include <arrows/mvg/kwiver_algo_mvg_export.h>
 
 #include <vital/types/feature_track_set.h>
 #include <vital/types/landmark_map.h>
@@ -52,7 +52,7 @@
 
 namespace kwiver {
 namespace arrows {
-namespace sfm {
+namespace mvg{
 
 
 /// Detect tracks which remain stationary in the image
@@ -67,7 +67,7 @@ namespace sfm {
  * \param [in] threshold  The threshold on pixel distance to the mean
  * \return                The set of stationary tracks
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 std::set<vital::track_sptr>
 detect_stationary_tracks(vital::feature_track_set_sptr tracks,
                          double threshold = 10.0);
@@ -86,7 +86,7 @@ detect_stationary_tracks(vital::feature_track_set_sptr tracks,
  *                        adjacent key frames and adjacent frames
  * \return                The set of selected frame numbers
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 std::set<vital::frame_id_t>
 keyframes_for_sfm(vital::feature_track_set_sptr tracks,
                   const vital::frame_id_t radius = 10,
@@ -106,7 +106,7 @@ typedef std::vector<coverage_pair> frame_coverage_vec;
  * \param [in] cams the set of frames to have coverage calculated on
  * \return     the image coverages
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 frame_coverage_vec
 image_coverages(
   std::vector<vital::track_sptr> const& trks,
@@ -125,7 +125,7 @@ typedef std::vector<std::unordered_set<vital::frame_id_t>> camera_components;
  * \return camera_components vector.  Each set in the vector represents a
  * different connected component.
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 camera_components
 connected_camera_components(
   vital::simple_camera_perspective_map::frame_to_T_sptr_map const& cams,
@@ -142,7 +142,7 @@ connected_camera_components(
  * \param [in] tracks  The set of tracks
  * \returns            A vector of all critical tracks
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 std::vector<vital::track_sptr>
 detect_critical_tracks(camera_components const& cc,
                        vital::feature_track_set_sptr tracks);
@@ -165,7 +165,7 @@ detect_critical_tracks(camera_components const& cc,
               to disable
 * \return set of landmark ids (track_ids) that were bad and should be removed
 */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 std::set<vital::landmark_id_t>
 detect_bad_landmarks(
   vital::simple_camera_perspective_map::frame_to_T_sptr_map const& cams,
@@ -182,7 +182,7 @@ detect_bad_landmarks(
  * \param [in] to_remove track ids for landmarks to set null
  * \param [in,out] lms landmark map to remove landmarks from
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 void
 remove_landmarks(const std::set<vital::track_id_t>& to_remove,
   vital::landmark_map::map_landmark_t& lms);
@@ -197,7 +197,7 @@ remove_landmarks(const std::set<vital::track_id_t>& to_remove,
  *             of coverage are included in the return set
  * \return set of frames that do not meet the coverage requirement
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 std::set<vital::frame_id_t>
 detect_bad_cameras(
   vital::simple_camera_perspective_map::frame_to_T_sptr_map const& cams,
@@ -225,7 +225,7 @@ detect_bad_cameras(
  * \param [in] min_landmark_inliers minimum number of inlier measurements to keep a landmark.
  *             Set to -1 to ignore.
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 void
 clean_cameras_and_landmarks(
   vital::simple_camera_perspective_map& cams,
@@ -247,7 +247,7 @@ clean_cameras_and_landmarks(
  * Up in the image is the negative Y-axis.  Up in the world defaults to
  * the postive Z-axis, but this is configurable by specifying \p up.
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 bool
 camera_upright(vital::camera_perspective const& camera,
                vital::vector_3d const& up = vital::vector_3d(0, 0, 1));
@@ -256,7 +256,7 @@ camera_upright(vital::camera_perspective const& camera,
 /*
  * \sa camera_upright
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 bool
 majority_upright(
   vital::camera_perspective_map::frame_to_T_sptr_map const& cameras,
@@ -275,7 +275,7 @@ cameras_above_plane(
  * with the ground normal vector aligned with the Z-axis.  It returns the
  * median location in X and Y, and a small percentile (5%) of the height in Z.
  */
-KWIVER_ALGO_SFM_EXPORT
+KWIVER_ALGO_MVG_EXPORT
 vital::vector_3d
 landmarks_ground_center(vital::landmark_map const& landmarks,
                         double ground_frac = 0.05);
